[Refactor] CategoryNode를 다른 파일로 분리 (#55) - 이제 더이상 CategoryNode는 CategoryTree와 동일한 파일에 정의되지 않음
