+++ src/contexts/category/CategoryContext.ts
@@ -1,5 +1,5 @@
 import React from 'react'
-import { CategoryTree } from 'datastructures/category/CategoryTree'
+import CategoryTree from 'datastructures/category/CategoryTree'
 
 interface CategoryContextProps {
   categoryTree: CategoryTree
